update typing, support react components
