+++ packages/react/types/styled.d.ts
@@ -1,4 +1,5 @@
-import { Globals, OnlyObject, OnlyNumber, OnlyString, OnlyStringNumeric, Properties } from './css'
+import { Globals, OnlyStringNumeric, Properties } from './css'
+import { ForwardRefComponent } from './polymorphic'
 import * as Default from './default'
 import * as Theme from './theme'
 
@@ -24,7 +25,8 @@ export type NarrowObject<T> = {
 	)
 }
 
-type UnionToIntersection<Union> = (
+/** Returns the given array of objects an an intersection of types. */
+type ArrayAsIntersection<T extends any[], Union = T[number]> = (
 	Union extends any
 		? (argument: Union) => void
 	: never
@@ -246,10 +248,10 @@ export interface CreatedCss<
 			<T extends OmitKey<
 				(
 					& {
-						[K in keyof UnionToIntersection<Args[number]>]?: Widen<keyof UnionToIntersection<Args[number]>[K]>
+						[K in keyof ArrayAsIntersection<Args>]?: Widen<keyof ArrayAsIntersection<Args>[K]>
 					}
 					& {
-						[K in Exclude<keyof T, keyof UnionToIntersection<Args[number]>>]: any
+						[K in Exclude<keyof T, keyof ArrayAsIntersection<Args>>]: any
 					}
 				),
 				'css'
@@ -315,31 +317,23 @@ export interface CreatedCss<
 					}
 				)
 			}
-		): /** Component */ (
-			<T extends OmitKey<
-				(
+		): /** Component */ ForwardRefComponent<
+			'span',
+			(
+				& (
 					& {
-						[K in keyof UnionToIntersection<Args[number]>]?: Widen<keyof UnionToIntersection<Args[number]>[K]>
+						[K in keyof ArrayAsIntersection<Args>]?: Widen<keyof ArrayAsIntersection<Args>[K]>
 					}
 					& {
-						[K in Exclude<keyof T, keyof UnionToIntersection<Args[number]>>]: any
-					}
-				),
-				'css'
-			>>(
-				props: (
-					& T
-					& {
-						/** Inline CSS. */
-						css?: Style<C>
+						[K in Exclude<number | string, keyof ArrayAsIntersection<Args>>]: any
 					}
 				)
-			) => {
-				className: string
-				selector: string
-				props: T
-			}
-		) & {
+				& {
+					/** Inline CSS. */
+					css?: Style<C>
+				}
+			)
+		> & {
 			className: string
 			selector: string
